pyproject.toml
[tool.poetry]
name = "slack-bot"
version = "0.1.0"
description = "first attempt to a slack bot"
authors = ["CubeVic <victoraguirre.f@gmail.com>"]
readme = "README.md"
packages = [{include = "slack_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
slack-sdk = "^3.26.1"
flask = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
